+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### 0.0.46 "Cape Town" (14.04.2015 - ...04.2015)
 * Cleaned 'test' directory
 * Fixed problem with tests
+* Fixed bug with localStorage DELETE FROM (without WHERE)
 
 ### 0.0.45 "Rio de Janeiro" (13.04.2015 - 13.04.2015)
 * Changed CRLF for alacon.js and alaserver.js to LF